debug finish

+++ src/ex/ex03.java
@@ -9,8 +9,8 @@ public class ex03 {
 		System.out.print("請問要幾筆資料? ");
 		int n = scn.nextInt();
 		int a = scn.nextInt();
-		int sum = 0;
-		int data [][] = new int [n][a];
+		float sum = 0;
+		float data [][] = new float [n][a];
 		for(int i=0;i<n;i++){
 			for(int j=0;j<a;j++){
 			data[i][j]=scn.nextInt();
@@ -21,13 +21,15 @@ public class ex03 {
 		System.out.println(var(sum,data,n,a)/(n*a));
 		System.out.println(std(var(sum,data,n,a)/(n*a)));
 	}
-	public static float var(int x,int data[][],int n,int a){
-		int y;
+	public static float var(float x,float data[][],int n,int a){
+		float y;
 		float sum=0;
 		for(int i=0;i<n;i++){
 			for(int j=0;j<a;j++){
-			y= (data[i][j])-x;
+			y=(float) (data[i][j])-x;
+			//System.out.println("y"+y);
 			sum=(float) (sum+(Math.pow(y, 2)));
+			//System.out.println("sum"+sum);
 			}
 		}
